Add ability to open generated changelog for editing.
Introduces a new option to be specified: `launchEditor`. When specified,
`release-it-lerna-changelog` will generate the changelog with
`lerna-changelog` (just like it has always done) then launch the
configured editor with a temporary file. This allows the person doing
the release to customize the changelog before continuing.

There are a few valid values for `launchEditor`:

* `false` - Disables the feature.
* `true` - The `process.env.EDITOR` value will be used
* any string - This string will be used as if it were a command. In
  order to interpolate the temporary file path in the string, you can
  use `${file}` in your configuration.