Show lerna-changelog output _before_ version prompt.
This is a significant refactor (of both implementation, and tests). The
main goal is to ensure that the changelog that is displayed before
selecting what the next version will be is from lerna-changelog (not
whatever is configured as `git.changelog` script).  This makes it much
much easier to see that lerna-changelog labels have been applied, and
(assuming they have) what the right version to select is.

This also significantly refactors the testing harness to much closer
simulate actually executing `lerna-changelog` and the various `git`
commands (`plugin.responses` is a simple matcher for the command and
returns the value).