Only run linting on node 14
This works around an issue with node/no-missing-import that doesn't
properly see that `node:path` is a valid module on Node 12.x.

They are using `is-core-module` to figure out if it is a module or not,
and it seems that `is-core-module` doesn't differentiate between core
modules that are available for CJS (the `node:` prefixed paths are only
available in 14.18+ and 16+ in CJS files) and those that are available
to ESM.

This is really just a bandaid, the right fix would be to send a PR to
is-core-module to allow it to check CJS vs MJS modules (since they are
actually different).

See https://github.com/mysticatea/eslint-plugin-node/issues/299